requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
asn1crypto==1.5.1
attrs==23.2.0
bcrypt==4.0.1
blinker==1.7.0
cachelib==0.12.0
cachetools==5.3.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloud-sql-python-connector==1.7.0
colorama==0.4.6
cryptography==42.0.5
dnspython==2.6.1
ecdsa==0.18.0
email-validator==2.1.0.post1
et-xmlfile==1.1.0
fastapi==0.109.2
Flask==3.0.2
Flask-Session==0.6.0
frozenlist==1.4.1
google-auth==2.28.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.3
httptools==0.6.1
httpx==0.26.0
idna==3.6
inflect==7.0.0
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
multidict==6.0.5
numpy==1.26.4
openpyxl==3.1.2
orjson==3.9.14
pandas==2.2.1
passlib==1.7.4
pg8000==1.30.4
psycopg2-binary==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.6.1
pydantic-extra-types==2.5.0
pydantic-settings==2.2.0
pydantic_core==2.16.2
python-dateutil==2.8.2
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
redis==5.0.3
requests==2.31.0
rsa==4.9
scramp==1.4.4
six==1.16.0
sniffio==1.3.0
sqlacodegen-v2==0.1.4
SQLAlchemy==2.0.27
starlette==0.36.3
typing_extensions==4.9.0
tzdata==2024.1
ujson==5.9.0
urllib3==2.2.1
uvicorn==0.27.1
watchfiles==0.21.0
websockets==12.0
Werkzeug==3.0.1
yarl==1.9.4
